+++ HISTORY.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 - Added publishing of git tag when publishin
-- Updated `README.md` by adding some documentation
+- Updated `README.md` by adding some documentation and examples
 - Fixed `.bumpversion.cfg` by making each version a subsection
 
 ## v0.1.0